requirements.txt
python-decouple~=3.8
langchain~=0.3.21
langchain-openai~=0.3.9
langchain-deepseek~=0.1.3
langchain-chroma~=0.2.2
langchain-community~=0.3.20
langchain-ollama~=0.3.0
langchain-core~=0.3.47
chromadb~=0.6.3
onnxruntime~=1.21.0
openai~=1.61.0
swarm~=0.1.0
dspy~=2.6.16
autogen-agentchat~=0.4.9.3
autogen-ext[openai,langchain]~=0.4.9.3

